Add functionality for creating report page

+++ wp-content/plugins/seoanalyzertool/seoanalyzertool.php
@@ -18,6 +18,9 @@ class SeoAnalyzerTool{
 
 		add_filter( 'page_template', array($this, 'fw_reserve_page_template'));
 
+		add_action('admin_post_submit-form', array($this, '_handle_form_action')); // If the user is logged in
+		add_action('admin_post_nopriv_submit-form', array($this, '_handle_form_action')); // If the user in not logged in
+
 		register_activation_hook(__FILE__, array($this,'seo_analyze_tools_plugin_activate')); //activate hook
 		register_deactivation_hook(__FILE__, array($this,'seo_analyze_tools_plugin_deactivate')); //deactivate hook
 
@@ -70,6 +73,10 @@ class SeoAnalyzerTool{
 
 			$page_template = dirname( __FILE__ ) . '/views/single-seo-analyze-tools.php';
 		}
+		if ( is_page( 'SEO Analyze Report' ) ) {
+
+			$page_template = dirname( __FILE__ ) . '/views/single-seo-analyze-report.php';
+		}
 		return $page_template;
 	}
 
@@ -81,18 +88,26 @@ class SeoAnalyzerTool{
 
 		$the_page_title = 'SEO Analyze Tools';
 		$the_page_name = 'SEO Analyze Tools';
-
+		$the_page_title_report = 'SEO Analyze Report';
+		$the_page_name_report = 'SEO Analyze Report';
 		// the menu entry...
 		delete_option("seo_analyze_tools_plugin_page_title");
-		add_option("seo_analyze_tools_plugin_page_title", $the_page_title, '', 'yes');
+		add_option("seo_analyze_tools_plugin_page_title", $the_page_title_report, '', 'yes');
+		delete_option("seo_analyze_report_plugin_page_title");
+		add_option("seo_analyze_report_plugin_page_title", $the_page_title_report, '', 'yes');
 		// the slug...
 		delete_option("seo_analyze_tools_plugin_page_name");
-		add_option("seo_analyze_tools_plugin_page_name", $the_page_name, '', 'yes');
+		add_option("seo_analyze_tools_plugin_page_name", $the_page_name_report, '', 'yes');
+		delete_option("seo_analyze_report_plugin_page_name");
+		add_option("seo_analyze_report_plugin_page_name", $the_page_name_report, '', 'yes');
 		// the id...
 		delete_option("seo_analyze_tools_plugin_page_id");
 		add_option("seo_analyze_tools_plugin_page_id", '0', '', 'yes');
+		delete_option("seo_analyze_report_plugin_page_id");
+		add_option("seo_analyze_report_plugin_page_id", '0', '', 'yes');
 
 		$the_page = get_page_by_title( $the_page_title );
+		$the_page_report = get_page_by_title( $the_page_title_report );
 
 		if ( ! $the_page ) {
 
@@ -109,8 +124,7 @@ class SeoAnalyzerTool{
 			// Insert the post into the database
 			$the_page_id = wp_insert_post( $_p );
 
-		}
-		else {
+		}else {
 			// the plugin may have been previously active and the page may just be trashed...
 
 			$the_page_id = $the_page->ID;
@@ -121,8 +135,34 @@ class SeoAnalyzerTool{
 
 		}
 
+	if( ! $the_page_report ){
+			// Create post object
+		$_p = array();
+		$_p['post_title'] = $the_page_title_report;
+		$_p['post_content'] = "[seo_analyze_report]";
+		$_p['post_status'] = 'publish';
+		$_p['post_type'] = 'page';
+		$_p['comment_status'] = 'closed';
+		$_p['ping_status'] = 'closed';
+		$_p['post_category'] = array(1); // the default 'Uncatrgorised'
+
+		// Insert the post into the database
+		$the_page_report_id = wp_insert_post( $_p );
+	}else {
+			// the plugin may have been previously active and the page may just be trashed...
+
+			$the_page_report_id = $the_page->ID;
+
+			//make sure the page is not trashed...
+			$the_page->post_status = 'publish';
+			$the_page_report_id = wp_update_post( $the_page_report );
+
+		}
+
 		delete_option( 'seo_analyze_tools_plugin_page_id' );
+		delete_option( 'seo_analyze_report_plugin_page_id' );
 		add_option( 'seo_analyze_tools_plugin_page_id', $the_page_id );
+		add_option( 'seo_analyze_report_plugin_page_id', $the_page_report_id );
 
 	}
 
@@ -134,11 +174,21 @@ class SeoAnalyzerTool{
 		$the_page_title = get_option( "seo_analyze_tools_plugin_page_title" );
 		$the_page_name = get_option( "seo_analyze_tools_plugin_page_name" );
 
+		$the_page_title_report = get_option( "seo_analyze_report_plugin_page_title" );
+		$the_page_name_report = get_option( "seo_analyze_report_plugin_page_name" );
+
 		//  the id of our page...
 		$the_page_id = get_option( 'seo_analyze_tools_plugin_page_id' );
+		$the_page_report_id = get_option( 'seo_analyze_report_plugin_page_id' );
 		if( $the_page_id ) {
 
 			wp_delete_post( $the_page_id, true); // this will trash, not delete
+
+		}
+
+		if( $the_page_report_id ) {
+
+			wp_delete_post( $the_page_report_id, true); // this will trash, not delete
 
 		}
 
@@ -146,6 +196,36 @@ class SeoAnalyzerTool{
 		delete_option("seo_analyze_tools_plugin_page_name");
 		delete_option("seo_analyze_tools_plugin_page_id");
 
+		delete_option("seo_analyze_report_plugin_page_title");
+		delete_option("seo_analyze_report_plugin_page_name");
+		delete_option("seo_analyze_report_plugin_page_id");
+
+	}
+
+//add_action('admin_post_submit-form', '_handle_form_action'); // If the user is logged in
+//add_action('admin_post_nopriv_submit-form', '_handle_form_action'); // If the user in not logged in
+	public function _handle_form_action(){
+		if(isset($_POST) && !empty($_POST['url'])){
+			$url = $_POST['url'];
+			$foo = false;
+
+			$_p = array();
+			$_p['post_title'] = $url;
+			$_p['post_content'] = "[seo_analyze_report]";
+			$_p['post_status'] = 'publish';
+			$_p['post_type'] = 'seoanalyzertool';
+			$_p['comment_status'] = 'closed';
+			$_p['ping_status'] = 'closed';
+			$_p['post_category'] = array(1); // the default 'Uncatrgorised'
+
+			// Insert the post into the database
+			$the_page_report_id = wp_insert_post( $_p );
+
+//			wp_redirect(get_home_url().'/seo-analyze-report');
+//			wp_redirect(get_permalink(get_post($the_page_report_id)));
+			require_once('single-seo-analyze-report.php');
+//			locate_template('Page SEO Analyze Report');
+		}
 	}
